logistics-route-optimizer/requirements.txt
mysql-connector-python==8.3.0
googlemaps==4.10.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
tabulate==0.9.0
